requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
langchain_community

